+++ tests/unit/learning/aqe-learning-engine.test.ts
@@ -1,9 +1,14 @@
 /**
- * Unit Tests for AQE Learning Engine
- * Unified learning engine with graceful degradation
+ * Unit Tests for AQE Learning Engine (core suite)
  *
- * Tests pattern storage, task routing, model recommendations,
- * and Claude Flow integration with fallbacks.
+ * Unified learning engine with graceful degradation. Covers initialization,
+ * status, routing (task + model), codebase analysis, guidance, and dispose.
+ * The heaviest sub-suites live in companion files (issue #448, step 2):
+ *
+ *   - aqe-learning-engine-patterns.test.ts — Pattern Learning + Task
+ *     Tracking + Experience Capture Integration.
+ *   - aqe-learning-engine-factory.test.ts — Factory functions +
+ *     DEFAULT_ENGINE_CONFIG (no `engine.initialize()` overhead).
  */
 
 import { describe, it, expect, beforeEach, afterEach, afterAll, vi } from 'vitest';
@@ -31,54 +36,9 @@ import {
   createDefaultLearningEngine,
   DEFAULT_ENGINE_CONFIG,
   type AQELearningEngineConfig,
-  type AQELearningEngineStatus,
 } from '../../../src/learning/aqe-learning-engine.js';
 import type { MemoryBackend, EventBus } from '../../../src/kernel/interfaces.js';
-import type { QEDomain } from '../../../src/learning/qe-patterns.js';
-import { ok, err } from '../../../src/shared/types/index.js';
-
-// ============================================================================
-// Mock Factory
-// ============================================================================
-
-function createMockMemoryBackend(): MemoryBackend {
-  const storage = new Map<string, unknown>();
-
-  return {
-    get: vi.fn((key: string) => Promise.resolve(storage.get(key) ?? null)),
-    set: vi.fn((key: string, value: unknown) => {
-      storage.set(key, value);
-      return Promise.resolve();
-    }),
-    delete: vi.fn((key: string) => {
-      storage.delete(key);
-      return Promise.resolve();
-    }),
-    has: vi.fn((key: string) => Promise.resolve(storage.has(key))),
-    keys: vi.fn(() => Promise.resolve(Array.from(storage.keys()))),
-    search: vi.fn((pattern: string) => {
-      const regex = new RegExp(pattern.replace(/\*/g, '.*'));
-      const matches = Array.from(storage.keys()).filter((k) => regex.test(k));
-      return Promise.resolve(matches);
-    }),
-    clear: vi.fn(() => {
-      storage.clear();
-      return Promise.resolve();
-    }),
-    size: vi.fn(() => Promise.resolve(storage.size)),
-    close: vi.fn(() => Promise.resolve()),
-    getState: vi.fn(() => ({ type: 'memory', ready: true })),
-  } as unknown as MemoryBackend;
-}
-
-function createMockEventBus(): EventBus {
-  return {
-    publish: vi.fn(),
-    subscribe: vi.fn().mockReturnValue(() => {}),
-    unsubscribe: vi.fn(),
-    emit: vi.fn(),
-  } as unknown as EventBus;
-}
+import { createMockMemoryBackend, createMockEventBus } from './_aqe-engine-test-helpers';
 
 // ============================================================================
 // Tests
@@ -196,107 +156,6 @@ describe('AQELearningEngine', () => {
     });
   });
 
-  describe('Pattern Learning', () => {
-    it('should store a pattern', async () => {
-      const result = await engine.storePattern({
-        patternType: 'test-template',
-        name: 'Unit Test Pattern',
-        description: 'Standard unit test pattern',
-        template: {
-          type: 'code',
-          content: 'describe("Test", () => { it("should work", () => {}); })',
-          variables: [],
-        },
-        context: {
-          tags: ['unit', 'typescript'],
-          framework: 'vitest',
-        },
-      });
-
-      expect(result.success).toBe(true);
-      expect(result.value?.id).toBeDefined();
-      expect(result.value?.patternType).toBe('test-template');
-    });
-
-    it('should search patterns', async () => {
-      // Store a pattern first
-      await engine.storePattern({
-        patternType: 'test-template',
-        name: 'UserService Test',
-        description: 'Testing user service',
-        template: {
-          type: 'code',
-          content: 'describe("UserService", () => {})',
-          variables: [],
-        },
-        context: { tags: ['user', 'service'] },
-      });
-
-      const results = await engine.searchPatterns('user service');
-
-      expect(results.success).toBe(true);
-    });
-
-    it('should get pattern by ID', async () => {
-      const storeResult = await engine.storePattern({
-        patternType: 'mock-pattern',
-        name: 'Database Mock',
-        description: 'Mock database connection',
-        template: {
-          type: 'code',
-          content: 'vi.mock("./db")',
-          variables: [],
-        },
-        context: { tags: ['mock', 'database'] },
-      });
-
-      const pattern = await engine.getPattern(storeResult.value!.id);
-
-      expect(pattern).toBeDefined();
-      expect(pattern?.name).toBe('Database Mock');
-    });
-
-    it('should return null for non-existent pattern', async () => {
-      const pattern = await engine.getPattern('non-existent-id');
-      expect(pattern).toBeNull();
-    });
-
-    it('should record pattern outcome', async () => {
-      const storeResult = await engine.storePattern({
-        patternType: 'test-template',
-        name: 'Outcome Test Pattern',
-        description: 'Pattern for outcome testing',
-        template: { type: 'code', content: 'test', variables: [] },
-        context: { tags: ['test'] },
-      });
-
-      const outcomeResult = await engine.recordOutcome({
-        patternId: storeResult.value!.id,
-        success: true,
-        metrics: { testsPassed: 10, testsFailed: 0 },
-      });
-
-      expect(outcomeResult.success).toBe(true);
-    });
-
-    it('should return error when not initialized', async () => {
-      const uninitEngine = createAQELearningEngine(memory, {
-        projectRoot: '/test',
-      });
-
-      const result = await uninitEngine.storePattern({
-        patternType: 'test-template',
-        name: 'Test',
-        description: 'Test',
-        template: { type: 'code', content: '', variables: [] },
-        context: { tags: [] },
-      });
-
-      expect(result.success).toBe(false);
-      expect(result.error?.message).toContain('not initialized');
-    });
-  });
-
   describe('Task Routing', () => {
     it('should route task to optimal agent', async () => {
       const result = await engine.routeTask({
@@ -380,66 +239,6 @@ describe('AQELearningEngine', () => {
     });
   });
 
-  describe('Task Tracking', () => {
-    it('should start and track a task', async () => {
-      const taskId = await engine.startTask(
-        'Generate tests for PaymentService',
-        'qe-test-architect',
-        'test-generation'
-      );
-
-      expect(taskId).toBeDefined();
-
-      const task = engine.getTask(taskId);
-      expect(task).toBeDefined();
-      expect(task?.task).toBe('Generate tests for PaymentService');
-      expect(task?.agent).toBe('qe-test-architect');
-    });
-
-    it('should record steps in task', async () => {
-      const taskId = await engine.startTask('Multi-step task');
-
-      await engine.recordStep(taskId, 'analyze', 'Found 5 files', 0.9);
-      await engine.recordStep(taskId, 'generate', 'Created tests', 0.85);
-
-      const task = engine.getTask(taskId);
-      expect(task?.steps).toHaveLength(2);
-      expect(task?.steps[0].action).toBe('analyze');
-      expect(task?.steps[1].quality).toBe(0.85);
-    });
-
-    it('should end task and return execution', async () => {
-      const taskId = await engine.startTask('Completable task');
-
-      await engine.recordStep(taskId, 'step1');
-      const execution = await engine.endTask(taskId, true, 'Great job!');
-
-      expect(execution).toBeDefined();
-      expect(execution?.task).toBe('Completable task');
-      expect(execution?.steps).toHaveLength(1);
-
-      // Task should be removed from active
-      const task = engine.getTask(taskId);
-      expect(task).toBeUndefined();
-    });
-
-    it('should track completed task count', async () => {
-      const taskId1 = await engine.startTask('Task 1');
-      const taskId2 = await engine.startTask('Task 2');
-
-      await engine.endTask(taskId1, true);
-      await engine.endTask(taskId2, true);
-
-      const stats = await engine.getStats();
-      expect(stats.completedTasks).toBe(2);
-    });
-
-    it('should handle ending non-existent task', async () => {
-      const result = await engine.endTask('non-existent-task', true);
-      expect(result).toBeUndefined();
-    });
-  });
-
   describe('Codebase Analysis', () => {
     it('should analyze codebase (local fallback)', async () => {
       const result = await engine.analyzeCodebase('/test/project', 'shallow');
@@ -504,55 +303,6 @@ describe('AQELearningEngine', () => {
     });
   });
 
-  describe('Experience Capture Integration', () => {
-    it('should capture experience on task completion', async () => {
-      const taskId = await engine.startTask(
-        'Experience capture task',
-        'qe-test-architect',
-        'test-generation'
-      );
-
-      await engine.recordStep(taskId, 'step1', 'Result', 0.9);
-      await engine.endTask(taskId, true);
-
-      // Experience should be captured (check via stats or event emission)
-      const stats = await engine.getStats();
-      expect(stats.experienceCapture.totalExperiences).toBeGreaterThanOrEqual(0);
-    });
-
-    it('should allow manual experience capture', () => {
-      const id = engine.startExperienceCapture('Manual capture task', {
-        agent: 'qe-test-architect',
-        domain: 'test-generation',
-        model: 'sonnet',
-      });
-
-      expect(id).toBeDefined();
-    });
-
-    it('should return experience capture service', () => {
-      const service = engine.getExperienceCaptureService();
-      expect(service).toBeDefined();
-    });
-
-    it('should work without experience capture enabled', async () => {
-      const noExpEngine = createAQELearningEngine(memory, {
-        projectRoot: '/test',
-        enableExperienceCapture: false,
-      });
-      await noExpEngine.initialize();
-
-      const taskId = await noExpEngine.startTask('No experience task');
-      await noExpEngine.endTask(taskId, true);
-
-      // Should not throw
-      const service = noExpEngine.getExperienceCaptureService();
-      expect(service).toBeUndefined();
-
-      await noExpEngine.dispose();
-    });
-  });
-
   describe('Dispose', () => {
     it('should dispose all components', async () => {
       const taskId = await engine.startTask('Pre-dispose task');
@@ -572,51 +322,5 @@ describe('AQELearningEngine', () => {
       await engine.dispose();
       await engine.dispose(); // Should not throw
     });
-  });
-});
-
-describe('Factory Functions', () => {
-  let memory: MemoryBackend;
-
-  beforeEach(() => {
-    memory = createMockMemoryBackend();
-  });
-
-  afterEach(() => {
-    vi.clearAllMocks();
-  });
-
-  it('should create engine with createAQELearningEngine', () => {
-    const engine = createAQELearningEngine(memory, {
-      projectRoot: '/test',
-    });
-
-    expect(engine).toBeInstanceOf(AQELearningEngine);
-  });
-
-  it('should create engine with createDefaultLearningEngine', () => {
-    const engine = createDefaultLearningEngine(memory, '/test');
-
-    expect(engine).toBeInstanceOf(AQELearningEngine);
-  });
-
-  it('should accept optional event bus', () => {
-    const eventBus = createMockEventBus();
-    const engine = createAQELearningEngine(
-      memory,
-      { projectRoot: '/test' },
-      eventBus
-    );
-
-    expect(engine).toBeInstanceOf(AQELearningEngine);
-  });
-});
-
-describe('DEFAULT_ENGINE_CONFIG', () => {
-  it('should have expected defaults', () => {
-    expect(DEFAULT_ENGINE_CONFIG.enableClaudeFlow).toBe(true);
-    expect(DEFAULT_ENGINE_CONFIG.enableExperienceCapture).toBe(true);
-    expect(DEFAULT_ENGINE_CONFIG.enablePatternPromotion).toBe(true);
-    expect(DEFAULT_ENGINE_CONFIG.promotionThreshold).toBe(3);
   });
 });
